Add minor refactorings and change application.properties to .yml
