+++ src/main/kotlin/de/leipzigtech/ba/BaApplication.kt
@@ -3,7 +3,7 @@ package de.leipzigtech.ba
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories
+
 
 
 @SpringBootApplication
