--- src/main/resources/application.properties
@@ -1,28 +1,0 @@
-spring.main.banner-mode=off
-logging.level.org.springframework=ERROR
-
-
-#spring.datasource.initialization-mode=always
-spring.datasource.platform=postgres
-spring.datasource.url=jdbc:postgresql://localhost:5432/le_tech
-spring.datasource.username=le_tech
-spring.datasource.password=ddk0Re!!wIjpGAPzfa.XrQ
-#spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
-
-# The SQL dialect makes Hibernate generate better SQL for the chosen database
-spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
-
-# Hibernate ddl auto (create, create-drop, validate, update)
-spring.jpa.hibernate.ddl-auto = update
-
-
-spring.jpa.generate-ddl=true
-spring.jpa.properties.javax.persistence.schema-generation.scripts.action=create
-spring.jpa.properties.javax.persistence.schema-generation.scripts.create-target=./src/main/resources/db/migration/create.sql
-spring.jpa.properties.javax.persistence.schema-generation.scripts.create-source=metadata
-spring.jpa.properties.hibernate.format_sql=true
-
-
-# oauth2 github
-spring.security.oauth2.client.registration.github.clientId=
-spring.security.oauth2.client.registration.github.clientSecret=
